app/requirements.txt
Flask==2.0.2
requests==2.27.1
python-dotenv==0.19.2
